금액 관련 사항 BigDecimal 데이터 타입 적용 Enum -> HashMap 구조 개편 및 명칭 변경 라이브러리 간소화 엔드포인트명 변경

+++ src/main/kotlin/com/ic/mockserver/card/api/request/CardApproveRequest.kt
@@ -1,6 +1,7 @@
 package com.ic.mockserver.card.api.request
 
 import jakarta.validation.constraints.NotBlank
+import java.math.BigDecimal
 
 
 data class CardApproveRequest(
@@ -8,5 +9,5 @@ data class CardApproveRequest(
     val cardNumber: String,
 
     @field:NotBlank
-    val amount: Int
+    val amount: BigDecimal
 )
